pyproject.toml
[tool.poetry]
name = "music-bot"
version = "0.1.0"
description = ""
authors = [
    "Theo Paris <theoparisdesigns@gmail.com>",
    "Jamie L <root@neon-is.fun>"
]
packages = [{ include = "bot" }]

[tool.poetry.dependencies]
python = ">=3.8"
"discord.py" = { extras = ["voice"], version = "^1.5.1" }
lazyConfig = "^0.5"
youtube_dl = "^2020.12.2"
youtube-search = "^1.1.1"

[tool.poetry.dev-dependencies]
black = "^20.8b1"

[tool.poetry.scripts]
start = "bot.init:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
